added products page hero section
